put the handeling conection part of the code seperate and added the filter on the method and assec to the correct information
