+++ Makefile
@@ -2,11 +2,11 @@
 #VARIABLES
 
 NAME 	:= webserv
-SRC 	:= main.cpp parser.cpp running_servers.cpp
+SRC 	:= main.cpp parser.cpp running_servers.cpp handle_connection.cpp
 SUBDIR	:= srcs/
 SRCS	:= $(addprefix $(SUBDIR),$(SRC))
 OBJ 	:= $(SRCS:.c=.o)
-CMP		:= g++
+CMP		:= c++
 FLAGS 	:= -Werror -Wall -Wextra -g -I lib
 
 #---------------------------------
@@ -33,7 +33,7 @@ RESET	:= \033[0m
 
 all : 
 #		 g++ -fsanitize=address -g src/main.cpp src/parser.cpp
-		g++ -std=c++11 -g src/error_responce.cpp src/running_servers.cpp src/main.cpp src/parser.cpp src/Socket.cpp
+		g++ -std=c++11 -g src/error_responce.cpp src/running_servers.cpp src/main.cpp src/parser.cpp src/Socket.cpp src/handle_connection.cpp
 
 # all	: $(NAME)
 # 		@echo "$(GREEN)Project built successfully !$(RESET)"
