+++ valid url.txt
@@ -1,0 +1,31 @@
+For blabla.com:
+
+    localhost:8080/blabla.com
+    localhost:8080/blabla.com/redirect // maybe (format is a bit different)
+
+For test.com:
+
+    localhost:8080/test.com
+    localhost:8080/test.com/redirect
+
+For yeaha.com:
+
+    localhost:8070/yeaha.com
+    localhost:8070/yeaha.com/uploads
+    localhost:8070/yeaha.com/files/photos
+    localhost:8070/yeaha.com/files/photos/main/hello/
+    localhost:8070/yeaha.com/.py
+
+For new.com:
+
+    localhost:8070/new.com
+    localhost:8070/new.com/files/photos
+    localhost:8070/new.com/files/photos/main/hello/
+    localhost:8070/new.com/.py
+
+For new2.com:
+
+    localhost:8070/new2.com
+    localhost:8070/new2.com/files/photos
+    localhost:8070/new2.com/files/photos/main/hello/
+    localhost:8070/new2.com/.py
